+++ test/resources/includes/include_3.kt
@@ -1,0 +1,1 @@
+fun include_3() = println("include_3")