+++ test/resources/includes/include_4.kt
@@ -1,0 +1,1 @@
+fun include_4() = println("include_4")