+++ test/resources/includes/rel_includes/include_1.kt
@@ -1,0 +1,1 @@
+fun include_1() = println("include_1")