+++ test/resources/includes/rel_includes/include_2.kt
@@ -1,0 +1,1 @@
+fun include_2() = println("include_2")